+++ web/models/Api/POST/cancel_booking.js
@@ -1,6 +1,6 @@
-const ApiError = require("../../Api/ApiError");
-const DataApi = require("../../Api/DataApi");
-const db = require('../db');
+const ApiError = require("../../../HandleAPI/ApiError");
+const DataApi = require("../../../HandleAPI/DataApi");
+const db = require('../../db');
 
 const cancel_booking = async (req, res, next) =>
 {
