+++ web/models/Api/POST/registration_worker.js
@@ -1,6 +1,6 @@
 const bcrypt = require("bcrypt");
-const ApiError = require("../../Api/ApiError");
-const DataApi = require("../../Api/DataApi");
+const ApiError = require("../../../HandleAPI/ApiError");
+const DataApi = require("../../../HandleAPI/DataApi");
 const db = require('../db');
 
 const registration_worker = async (req, res, next) =>
@@ -30,21 +30,31 @@ const registration_worker = async (req, res, next) =>
             if (err)
                 return next(ApiError.internal('Internal error with hash password'));
 
-            // await db.query('CALL add_worker($1,$2,$3,$4,$5,$6,$7,$8, $9, $10,$11)', [
-            //     login,
-            //     hash,
-            //     password,
-            //     job_role,
-            //     surname,
-            //     first_name,
-            //     salary,
-            //     patronymic,
-            //     email,
-            //     phone,
+            db.query('CALL add_worker($1,$2,$3,$4,$5,$6,$7,$8, $9, $10,$11)', [
+                login,
+                hash,
+                password,
+                job_role,
+                surname,
+                first_name,
+                salary,
+                patronymic,
+                email,
+                phone,
+                job_rate
+            ])
+                .then((data) =>
+                {
+                    console.log(data);
+                    return next(DataApi.success({}, "Request execution"));
+                })
+                .catch((err) =>
+                {
+                    console.log("Error with registration worker", { color: "red" });
+                    console.error(err);
+                    return next(ApiError.badRequest("User is exists"));
+                })
 
-            //     job_rate
-            // ]);
-            return next(DataApi.success({}, "Request execution"));
         })
     });
